int to float (again) - test

+++ lib/rapuma/core/page_background.py
@@ -178,8 +178,8 @@ class PageBackground (object) :
 
         # input and calculation
         # get the values for page dimensions from Rapuma usfm_layout.conf 
-        pageHeight          = int(self.layoutConfig['PageLayout']['pageHeight'])
-        pageWidth           = int(self.layoutConfig['PageLayout']['pageWidth'])
+        pageHeight          = float(self.layoutConfig['PageLayout']['pageHeight'])
+        pageWidth           = float(self.layoutConfig['PageLayout']['pageWidth'])
 
         # The page size is defined in [mm] but is entered in pixels [px] 
         # The conversion factor for [px] is 90/25.4
